requirements.txt
av==12.3.0
certifi==2024.8.30
charset-normalizer==3.3.2
coloredlogs==15.0.1
ctranslate2==4.4.0
decorator==4.4.2
faster-whisper==1.0.3
ffmpeg-python==0.2.0
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.9.0
future==1.0.0
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
openai-whisper==20240930
packaging==24.1
pillow==10.4.0
proglog==0.1.10
protobuf==5.28.2
pytube==15.0.0
pytubefix==7.2.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
setuptools==75.1.0
sympy==1.13.3
tiktoken==0.7.0
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
